work on some better abstracitons?
