+++ api/src/main/kotlin/dev/jaims/moducore/api/data/Kit.kt
@@ -36,35 +36,29 @@ data class Kit(
     var consoleCommands: List<String>,
     var playerCommands: List<String>,
     val kitInfo: KitInfo = KitInfo(name, "Your Custom Kit Description! Edit in kits.yml", "DIRT", false)
-) {
-    fun give(player: Player): Kit {
-        // give items
-        val extras = player.inventory.addItem(*items.toTypedArray())
-        extras.forEach { (_, item) ->
-            player.world.dropItem(player.location, item)
-        }
-        // run player commands
-        playerCommands.forEach { command ->
-            val success = player.performCommand(command.replace("{player}", player.name))
-            if (!success) {
-                Errors.KIT_PLAYER_COMMAND_FAILED.log(command, player.name, name)
-            }
-        }
-        // run console commands
-        consoleCommands.forEach { command ->
-            val success = Bukkit.getServer()
-                .dispatchCommand(Bukkit.getServer().consoleSender, command.replace("{player}", player.name))
-            if (!success) {
-                Errors.KIT_CONSOLE_COMMAND_FAILED.log(command, player.name, name)
-            }
-        }
-        return this
-    }
-}
-
-data class KitInfo(
-    val displayName: String,
-    val description: String,
-    val displayItem: String,
-    val glow: Boolean
 )
+
+
+fun Kit.give(player: Player): Kit {
+    // give items
+    val extras = player.inventory.addItem(*items.toTypedArray())
+    extras.forEach { (_, item) ->
+        player.world.dropItem(player.location, item)
+    }
+    // run player commands
+    playerCommands.forEach { command ->
+        val success = player.performCommand(command.replace("{player}", player.name))
+        if (!success) {
+            Errors.KIT_PLAYER_COMMAND_FAILED.log(command, player.name, name)
+        }
+    }
+    // run console commands
+    consoleCommands.forEach { command ->
+        val success = Bukkit.getServer()
+            .dispatchCommand(Bukkit.getServer().consoleSender, command.replace("{player}", player.name))
+        if (!success) {
+            Errors.KIT_CONSOLE_COMMAND_FAILED.log(command, player.name, name)
+        }
+    }
+    return this
+}
