+++ api/src/main/kotlin/dev/jaims/moducore/api/data/KitInfo.kt
@@ -1,0 +1,8 @@
+package dev.jaims.moducore.api.data
+
+data class KitInfo(
+    val displayName: String,
+    val description: String,
+    val displayItem: String,
+    val glow: Boolean
+)
