+++ api/src/main/kotlin/dev/jaims/moducore/api/manager/Reloadable.kt
@@ -1,0 +1,10 @@
+package dev.jaims.moducore.api.manager
+
+interface Reloadable {
+
+    /**
+     * Reload this [Reloadable]
+     */
+    fun reload()
+
+}
